Windows support.

+++ src/system.cpp
@@ -6,9 +6,10 @@
 #include <cassert>
 #include <cstdint>
 #include <cstdio>
+#include <cstdlib>
 #include <cstring>
 
-#include <time.h>
+#include "compat/time.hpp"
 
 #include "system.hpp"
 
@@ -139,7 +140,7 @@ namespace c8e
                     case 0x0004: // 0x8XY4 : Adds VY to VX. VF is set to 1 when there's a carry, and to 0 when there isn't.
                     {
                         const uint16_t res = sys->V[regX] + sys->V[regY];
-                        sys->V[regX] = res;
+                        sys->V[regX] = uint8_t(res);
                         sys->VF      = (res >> 8 ? 1 : 0);
                         break;
                     }
@@ -147,7 +148,7 @@ namespace c8e
                     case 0x0005: // 0x8XY5 : VY is subtracted from VX. VF is set to 0 when there's a borrow, and 1 when there isn't.
                     {
                         const uint16_t res = sys->V[regX] - sys->V[regY];
-                        sys->V[regX] = res;
+                        sys->V[regX] = uint8_t(res);
                         sys->VF      = (res >> 8 ? 0 : 1);
                         break;
                     }
@@ -160,7 +161,7 @@ namespace c8e
                     case 0x0007: // 0x8XY7 : Sets VX to VY minus VX. VF is set to 0 when there's a borrow, and 1 when there isn't.
                     {
                         const uint16_t res = sys->V[regY] - sys->V[regX];
-                        sys->V[regX] = res;
+                        sys->V[regX] = uint8_t(res);
                         sys->VF      = (res >> 8 ? 0 : 1);
                         break;
                     }
